requirements.txt
# Core MCP Server dependencies
mcp>=0.9.0
# asyncio is built into Python 3.7+
pydantic>=2.0.0
click>=8.1.0

# HTTP and networking
aiohttp>=3.8.0
httpx>=0.25.0
requests>=2.28.0

# DNS and network utilities
dnspython>=2.3.0
python-nmap>=0.7.1

# Web scraping and parsing
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.15.0

# Security and cryptography
cryptography>=41.0.0
pyjwt>=2.8.0
pyotp>=2.9.0

# Database and data parsing
sqlparse>=0.4.4

# API integrations
shodan>=1.28.0
censys>=2.2.0

# Utility libraries
colorama>=0.4.6
tabulate>=0.9.0
tqdm>=4.66.0
python-dotenv>=1.0.0
qrcode>=7.4.0
validators>=0.22.0
fake-useragent>=1.4.0
cloudscraper>=1.2.71

# System utilities
psutil>=5.9.0
paramiko>=3.3.0
asyncssh>=2.14.0

# Optional dependencies for advanced features
# Uncomment as needed:

# For PDF report generation
reportlab>=4.0.0
weasyprint>=60.0

# For image processing and screenshots
Pillow>=10.0.0
playwright>=1.40.0

# For advanced data analysis
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.7.0

# For machine learning-based detection
scikit-learn>=1.3.0
tensorflow>=2.13.0

# For YAML configuration support
PyYAML>=6.0

# For advanced wordlist management
python-magic>=0.4.27

# For container scanning
# docker>=6.1.0

# For cloud security
# boto3>=1.29.0  # AWS
# azure-identity>=1.14.0  # Azure
# google-cloud-core>=2.3.0  # GCP
